/ ‘Project/my_code/calc.C’

+++ Project/my_code/calc.C
@@ -19,7 +19,7 @@ main()
 		iStream >> num;
 		total += num;
 	}
-	cout << total/100 << (total/100)/250 << endl;
+	cout << total/100 << ((float)total/100)/250 << endl;
 
 	total = 0;
 	for (int i = 0; i != 100; ++i) {
@@ -28,7 +28,7 @@ main()
 		iStream >> num;
 		total += num;
 	}
-	cout << total/100 << (total/100)/250 << endl;
+	cout << total/100 << ((float)total/100)/250 << endl;
 
 	return 0;
 }
